+++ crates/rspack_core/src/tree_shaking/optimizer.rs
@@ -31,7 +31,7 @@ use super::{
 };
 use crate::{
   contextify, dbg_matches, join_string_component, tree_shaking::ConvertModulePath, Compilation,
-  ModuleGraph, ModuleIdentifier, ModuleSyntax, ModuleType, NormalModuleAstOrSource,
+  DependencyType, ModuleGraph, ModuleIdentifier, ModuleSyntax, ModuleType, NormalModuleAstOrSource,
 };
 
 pub struct CodeSizeOptimizer<'a> {
@@ -39,6 +39,11 @@ pub struct CodeSizeOptimizer<'a> {
   bailout_modules: IdentifierMap<BailoutFlag>,
   side_effects_free_modules: IdentifierSet,
   symbol_graph: SymbolGraph,
+}
+
+enum EntryLikeType {
+  Entry,
+  Bailout,
 }
 
 impl<'a> CodeSizeOptimizer<'a> {
@@ -156,14 +161,14 @@ impl<'a> CodeSizeOptimizer<'a> {
       visited_symbol_ref,
       &dead_nodes_index,
     );
-    // dbg!(&used_symbol_ref);
-    let used_modules_ids = self
-      .compilation
-      .module_graph
-      .module_identifier_to_module_graph_module
-      .iter()
-      .map(|(id, mgm)| (*id, mgm.used))
-      .collect::<Vec<_>>();
+    dbg!(&used_symbol_ref);
+    // let used_modules_ids = self
+    //   .compilation
+    //   .module_graph
+    //   .module_identifier_to_module_graph_module
+    //   .iter()
+    //   .map(|(id, mgm)| (*id, mgm.used))
+    //   .collect::<Vec<_>>();
     // dbg!(&used_modules_ids);
     Ok(
       OptimizeDependencyResult {
@@ -263,23 +268,7 @@ impl<'a> CodeSizeOptimizer<'a> {
     let bailout_entry_modules = self
       .bailout_modules
       .iter()
-      .filter_map(|(module_id, reason)| {
-        if reason
-          .intersection(
-            BailoutFlag::DYNAMIC_IMPORT
-              | BailoutFlag::HELPER
-              | BailoutFlag::COMMONJS_REQUIRE
-              | BailoutFlag::CONTEXT_MODULE,
-          )
-          .bits()
-          .count_ones()
-          >= 1
-        {
-          Some(*module_id)
-        } else {
-          None
-        }
-      })
+      .map(|(module_id, reason)| *module_id)
       .collect::<Vec<_>>();
     for module_id in bailout_entry_modules {
       self.collect_from_entry_like(
@@ -289,7 +278,7 @@ impl<'a> CodeSizeOptimizer<'a> {
         used_export_module_identifiers,
         &inherit_export_ref_graph,
         &mut traced_tuple,
-        false,
+        EntryLikeType::Bailout,
         visited_symbol_ref,
         errors,
       );
@@ -315,7 +304,7 @@ impl<'a> CodeSizeOptimizer<'a> {
         used_export_module_identifiers,
         inherit_export_ref_graph,
         traced_tuple,
-        true,
+        EntryLikeType::Entry,
         visited_symbol_ref,
         errors,
       );
@@ -360,9 +349,9 @@ impl<'a> CodeSizeOptimizer<'a> {
           .compilation
           .entry_modules()
           .chain(context_entry_modules)
-          .map(|module_id| (module_id, true)),
+          .map(|module_id| module_id),
       );
-      while let Some((module_identifier, _is_entry)) = q.pop_front() {
+      while let Some(module_identifier) = q.pop_front() {
         if visited.contains(&module_identifier) {
           continue;
         } else {
@@ -436,11 +425,11 @@ impl<'a> CodeSizeOptimizer<'a> {
             panic!("Failed to get ModuleGraphModule by module identifier {module_identifier}")
           });
         // reachable_dependency_identifier.extend(analyze_result.inherit_export_maps.keys());
-        for dep in mgm.dependencies.iter() {
+        for dependency_id in mgm.dependencies.iter() {
           let module_ident = match self
             .compilation
             .module_graph
-            .module_identifier_by_dependency_id(dep)
+            .module_identifier_by_dependency_id(dependency_id)
           {
             Some(module_identifier) => module_identifier,
             None => {
@@ -451,13 +440,12 @@ impl<'a> CodeSizeOptimizer<'a> {
                 .and_then(|module| module.as_normal_module())
                 .map(|normal_module| normal_module.ast_or_source())
               {
-                Some(ast_or_source) => {
-                  if matches!(ast_or_source, NormalModuleAstOrSource::BuiltFailed(_)) {
-                    // We know that the build output can't run, so it is alright to generate a wrong tree-shaking result.
-                    continue;
-                  } else {
-                    panic!("Failed to resolve {dep:?}")
-                  }
+                Some(NormalModuleAstOrSource::BuiltFailed(_)) => {
+                  // We know that the build output can't run, so it is alright to generate a wrong tree-shaking result.
+                  continue;
+                }
+                Some(_) => {
+                  panic!("Failed to ast of {dependency_id:?}")
                 }
                 None => {
                   panic!("Failed to get normal module of {module_identifier}");
@@ -465,13 +453,33 @@ impl<'a> CodeSizeOptimizer<'a> {
               };
             }
           };
+          let dependency = match self
+            .compilation
+            .module_graph
+            .dependency_by_id(dependency_id)
+          {
+            Some(dep) => dep,
+            None => {
+              // It means this dependency has been removed before
+              continue;
+            }
+          };
+
+          let need_bailout = matches!(
+            dependency.dependency_type(),
+            DependencyType::CommonJSRequireContext
+              | DependencyType::RequireContext
+              | DependencyType::DynamicImport
+              | DependencyType::CjsRequire
+              | DependencyType::ImportContext
+          );
           if self.side_effects_free_modules.contains(module_ident)
             && !reachable_dependency_identifier.contains(module_ident)
-            && !self.bailout_modules.contains_key(module_ident)
+            && !need_bailout
           {
             continue;
           }
-          q.push_back((*module_ident, false));
+          q.push_back(*module_ident);
         }
         // dbg!(&module_identifier);
         // dbg!(&reachable_dependency_identifier);
@@ -1055,7 +1063,7 @@ impl<'a> CodeSizeOptimizer<'a> {
     used_export_module_identifiers: &mut IdentifierMap<ModuleUsedType>,
     inherit_extend_graph: &GraphMap<ModuleIdentifier, (), Directed>,
     traced_tuple: &mut HashMap<(ModuleIdentifier, ModuleIdentifier), Vec<(SymbolRef, SymbolRef)>>,
-    is_entry: bool,
+    entry_type: EntryLikeType,
     visited_symbol_ref: &mut HashSet<SymbolRef>,
     errors: &mut Vec<Error>,
   ) {
@@ -1070,10 +1078,11 @@ impl<'a> CodeSizeOptimizer<'a> {
     };
 
     // by default webpack will not mark the `export *` as used in entry module
-    if !is_entry {
+    if matches!(entry_type, EntryLikeType::Bailout) {
       let inherit_export_symbols = get_inherit_export_symbol_ref(entry_module_result);
 
       q.extend(inherit_export_symbols);
+      q.extend(entry_module_result.used_symbol_refs.iter().cloned());
     }
 
     for item in entry_module_result.export_map.values() {
@@ -1090,9 +1099,9 @@ impl<'a> CodeSizeOptimizer<'a> {
       );
     }
 
-    while let Some(sym_ref) = q.pop_front() {
+    while let Some(symbol_ref) = q.pop_front() {
       self.mark_symbol(
-        sym_ref,
+        symbol_ref,
         analyze_map,
         &mut q,
         evaluated_module_identifiers,
@@ -1236,7 +1245,7 @@ async fn par_analyze_module(
         // ast.visit(|p, c| {
         //   dbg!(&p);
         // });
-        dbg_matches!(
+        dbg!(
           *uri_key,
           &uri_key,
           // &analyzer.export_all_list,
@@ -1246,7 +1255,7 @@ async fn par_analyze_module(
           &analyzer.used_symbol_refs,
           analyzer.helper_mark,
           analyzer.top_level_mark,
-          analyzer.unresolved_mark
+          analyzer.unresolved_mark,
         );
 
         Some((uri_key, analyzer))
