+++ examples/basic/source/package.json
@@ -1,3 +1,3 @@
 {
-  "sideEffects": true
+  "sideEffects": false
 }
